#338 Update to learning form
